extract GetAssetKey patch into DmProRendererContext
