+++ SolastaUnfinishedBusiness/Models/DmProRendererContext.cs
@@ -251,4 +251,42 @@ internal static class DmProRendererContext
             reflectionProbe.transform.position = new Vector3(size.x / 2f, size.y, size.z / 2f);
         }
     }
+    
+    internal static bool ExtendedGetAssetKey(
+        BaseBlueprint __instance,
+        ref string __result,
+        BaseBlueprint.PrefabByEnvironmentDescription prefabByEnvironmentDescription,
+        EnvironmentDefinition environmentDefinition,
+        bool perspective)
+    {
+        if (__instance is not PropBlueprint propBlueprint || !propBlueprint.Name.EndsWith("MOD"))
+        {
+            return true;
+        }
+
+        //PATCH: ensures custom props display the proper icon (DMP)
+        var a = propBlueprint.Name.Split(new[] { '~' }, 3);
+
+        if (a.Length != 3)
+        {
+            return true;
+        }
+
+        var propName = a[0];
+        var environmentName = a[1];
+        var str1 = "Gui/Bitmaps/Blueprints/Props/";
+        var str2 = "User_Props_" + propName;
+        var postfix = perspective ? "_Pers" : "_Top";
+
+        if (environmentDefinition != null &&
+            prefabByEnvironmentDescription.Environment == environmentDefinition.Name)
+        {
+            str1 = "Gui/Bitmaps/Props-" + environmentName + "/";
+            str2 = str2 + "_" + environmentName;
+        }
+
+        __result = str1 + str2 + postfix;
+
+        return false;
+    }
 }
